requirements.txt
accelerate==0.30.1
certifi==2024.2.2
charset-normalizer==3.3.2
diffusers==0.27.2
filelock==3.14.0
fsspec==2024.5.0
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
peft==0.11.1
pillow==10.3.0
psutil==5.9.8
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.0
safetensors==0.4.3
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.11.0
urllib3==2.2.1
zipp==3.18.2
